Add Exp for asm

+++ jvm/src/main/java/example/main.java
@@ -2,7 +2,7 @@ package example;
 
 public class main {
     public static void main(String[] args) {
-        Object result = new add().invoke(new Integer(1), new Integer(2));
+        Object result = new add().invoke(new Double(1), new Double(2));
         System.out.println(result);
     }
 }
