+++ jvm/src/main/java/kand/runtime/Core.java
@@ -2,6 +2,6 @@ package kand.runtime;
 
 public class Core {
     public static Object add(Object x, Object y) {
-        return new Integer((Integer)x + (Integer)y);
+        return new Double((Double)x + (Double)y);
     }
 }
